refactor(analyzer-rust): modularize fastify analyzer internals
